@@ -1,6 +1,6 @@
 {
   "name": "gitbook-styleguide",
-  "version": "7.0.0-pre.8",
+  "version": "7.9.0",
   "main": "./index.js",
   "author": "GitBook <contact@gitbook.com>",
   "homepage": "http://styleguide.gitbook.com/",
@@ -9,7 +9,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "classnames": "^2.2.5",
-    "gitbook-markdown-css": "1.0.1",
+    "gitbook-markdown-css": "1.1.0",
     "moment": "^2.14.1",
     "moment-duration-format": "^1.3.0",
     "octicons": "3.4.1",
@@ -27,7 +27,7 @@
     "eslint": "^2.11.1",
     "eslint-config-gitbook": "1.2.0",
     "less": "^2.7.1",
-    "next": "1.1.1",
+    "next": "2.0.0-beta.13",
     "parallelshell": "^2.0.0",
     "slate": "^0.16.8",
     "slate-edit-code": "^0.10.0",
