@@ -5,13 +5,25 @@ const Container = React.createClass({
     propTypes: {
         className: React.PropTypes.string,
         children:  React.PropTypes.node,
-        fluid:     React.PropTypes.bool
+        // No hoz padding
+        fluid:     React.PropTypes.bool,
+        // Take all the width
+        maxWidth:  React.PropTypes.bool
+    },
+
+    getDefaultProps() {
+        return {
+            maxWidth: false,
+            fluid: false
+        };
     },
 
     render() {
-        let { fluid, className, ...props } = this.props;
+        let { fluid, className, maxWidth, ...props } = this.props;
+
         className = classNames('container', {
-            'container-fluid': fluid
+            'container-fluid': fluid,
+            'container-max-width': maxWidth
         }, className);
 
         return (
